requirements.txt
async-generator==1.10
attrs==22.1.0
autopep8==1.7.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
chromedriver-binary==107.0.5304.18.0
colorama==0.4.5
docopt==0.6.2
exceptiongroup==1.0.0rc9
h11==0.14.0
idna==3.4
MouseInfo==0.1.3
oauthlib==3.2.1
outcome==1.2.0
Pillow==9.2.0
pipreqs==0.4.11
PyAutoGUI==0.9.53
pycodestyle==2.9.1
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-dotenv==0.21.0
pytweening==1.0.4
requests==2.28.1
requests-oauthlib==1.3.1
selenium==4.5.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
toml==0.10.2
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
tweepy==4.10.1
urllib3==1.26.12
webdriver-manager==3.8.3
wsproto==1.2.0
yarg==0.1.9
